requirements.txt
pyodbc
markdown
zulip
kafka-python
